Convert test repo into recipe project repo
The friedRice repository was just a test repository I created playing with git commands
in Bash, however now changing it to the recipe project for a simple website to display
a couple of my favorite meals.

Deleted many of the test files, and editted the README to match the new project.