Created basic HTML structure with content
Established basic HTML structure for a simple web
page layout. A simple header with links to other
pages for recipes. A main body including a descri-
ption, image and steps for the recipe.
An aside that shows the ingredients for the recipe.
And a footer for the author of the page and for
anything else I can think of later...